fix(trpc): resolve the owner's relay for run-now and share the resolver
runNow (the manual 'run automation now' trigger) hardcoded env.RELAY_URL,
the same gap as scheduled dispatch — a host on an overridden relay was
unreachable. Extract resolveUserRelayUrl and route relayEndpoint, runNow,
and dispatch through it.
