+++ packages/trpc/src/router/automation/automation.ts
@@ -15,7 +15,7 @@ import {
 import { TRPCError, type TRPCRouterRecord } from "@trpc/server";
 import { and, desc, eq, getTableColumns, ilike } from "drizzle-orm";
 import { z } from "zod";
-import { env } from "../../env";
+import { resolveUserRelayUrl } from "../../lib/relay-url";
 import { protectedProcedure } from "../../trpc";
 import { requireActiveOrgMembership } from "../utils/active-org";
 import { dispatchAutomation } from "./dispatch";
@@ -565,7 +565,7 @@ export const automationRouter = {
 			const outcome = await dispatchAutomation({
 				automation,
 				scheduledFor: new Date(),
-				relayUrl: env.RELAY_URL,
+				relayUrl: await resolveUserRelayUrl(automation.ownerUserId),
 			});
 
 			if (outcome.status === "conflict") {
