+++ packages/trpc/src/router/host/host.ts
@@ -11,18 +11,12 @@ import {
 	isActiveSubscriptionStatus,
 	isPaidPlan,
 } from "@superset/shared/billing";
-import { FEATURE_FLAGS } from "@superset/shared/constants";
 import { parseHostRoutingKey } from "@superset/shared/host-routing";
 import { TRPCError, type TRPCRouterRecord } from "@trpc/server";
 import { and, desc, eq, inArray } from "drizzle-orm";
 import { z } from "zod";
-import { env } from "../../env";
-import { posthog } from "../../lib/analytics";
+import { resolveUserRelayUrl } from "../../lib/relay-url";
 import { jwtProcedure, protectedProcedure } from "../../trpc";
-
-interface RelayUrlPayload {
-	url?: string;
-}
 
 export const hostRouter = {
 	/**
@@ -32,19 +26,7 @@ export const hostRouter = {
 	 * silently fell back, which split hosts and clients across two relays.
 	 */
 	relayEndpoint: jwtProcedure.query(async ({ ctx }) => {
-		try {
-			const payload = (await posthog.getFeatureFlagPayload(
-				FEATURE_FLAGS.RELAY_URL_OVERRIDE,
-				ctx.userId,
-			)) as RelayUrlPayload | null | undefined;
-			const override = payload?.url;
-			if (typeof override === "string" && override.length > 0) {
-				return { url: override };
-			}
-		} catch {
-			// Fall through to the default relay.
-		}
-		return { url: env.RELAY_URL };
+		return { url: await resolveUserRelayUrl(ctx.userId) };
 	}),
 
 	list: jwtProcedure
